Break Rosenbrock helper classes into own files
