+++ GridKit/Solver/Dynamic/Rosenbrock.cpp
@@ -944,85 +944,5 @@ namespace Integrator
     return 0;
   }
 
-  /**
-   * @brief Standard textbook adaptive controller. Accept if `err <= 1` and use
-   *
-   * \f[h_{new} = h * \min \left\{fac_{max}, \max\left\{fac_{min}, fac_{scale} \cdot e ^{-1/p}\right\}\right\}.\f]
-   *
-   */
-  template <typename RealT>
-  StepControl<RealT> AdaptiveStep<RealT>::nextStep(RealT err, StepControl<RealT> prev_step, uint8_t method_order)
-  {
-    StepControl<RealT> next_step = prev_step;
-
-    double h_mult = std::min(params_.fac_max_, std::max(params_.fac_scale_ * std::pow(err, GridKit::MINUS_ONE<RealT> / method_order), params_.fac_min_));
-
-    next_step.accept_     = err <= 1;
-    next_step.step_size_ *= h_mult;
-
-    return next_step;
-  }
-
-  /**
-   * @brief Fixed step - accept every step, no matter the error, and keep the step size the same.
-   *
-   */
-  template <typename RealT>
-  StepControl<RealT> FixedStep<RealT>::nextStep([[maybe_unused]] RealT err, StepControl<RealT> prev_step, [[maybe_unused]] uint8_t method_order)
-  {
-    return StepControl<RealT>{
-        .accept_    = true,
-        .step_size_ = prev_step.step_size_,
-    };
-  }
-
-  /**
-   * @brief Calculate the infinity error norm as
-   *
-   * \f[e = \max_i\frac{|\hat{e}_i|}{Atol_i + Rtol \cdot \max\{|y_{0i}|, |y_{1i}|\}},\f]
-   *
-   * where \f(y_0\f) is the initial state, \f(y_1\f) is the next state, and \f(\hat{e}\f) is the estimated error made in calculating
-   * the next state (typically \f(\hat{e} = y_1 - \hat{y}_1\f) for some different-order approximation \f(\hat{y}_1\f)).
-   *
-   * @param err \f(\hat{e}\f) in the above formula.
-   * @param y \f(y_1\f) in the above formula.
-   * @param yprev \f(y_0\f) in the above formula.
-   * @param handler The handler to be used for performing linear algebra operations.
-   * @param memspace The memory space to be used for performing linear lagebra operations.
-   * @see `Rosenbrock::errorEstimate()`
-   */
-  template <class ScalarT, typename IdxT>
-  InfNorm<ScalarT, IdxT>::RealT InfNorm<ScalarT, IdxT>::errorNorm(State& err, State& y, State& yprev, GridKit::LinearAlgebra::VectorHandler<ScalarT, IdxT>& handler, GridKit::memory::MemorySpace memspace) const
-  {
-    if (int err_code = workspace_.out_->copyFromExternal(&err, memspace, memspace))
-    {
-      throw std::format("GridKit::LinearAlgebra::Vector::copyFromExternal failed with error code {}", err_code);
-    }
-    if (int err_code = workspace_.scale_->copyFromExternal(&y, memspace, memspace))
-    {
-      throw std::format("GridKit::LinearAlgebra::Vector::copyFromExternal failed with error code {}", err_code);
-    }
-    if (int err_code = workspace_.yprev_abs_->copyFromExternal(&yprev, memspace, memspace))
-    {
-      throw std::format("GridKit::LinearAlgebra::Vector::copyFromExternal failed with error code {}", err_code);
-    }
-
-    handler.abs(workspace_.scale_.get(), workspace_.scale_.get(), memspace);
-    handler.abs(workspace_.yprev_abs_.get(), workspace_.scale_.get(), memspace);
-    handler.max(workspace_.yprev_abs_.get(), workspace_.scale_.get(), workspace_.scale_.get(), memspace);
-
-    // TODO: This scal shouldn't be necessary, but axpy doesn't support scaling the y parameter. In the future,
-    // the scaling should be able to be put on the next axpy.
-    handler.scal(params_.rel_tol_, workspace_.scale_.get(), memspace);
-    handler.axpy(GridKit::ONE<RealT>, params_.abs_tol_.get(), workspace_.scale_.get(), memspace);
-    handler.diagSolve(workspace_.scale_.get(), workspace_.out_.get(), memspace);
-
-    return handler.amax(workspace_.out_.get(), memspace);
-  }
-
   template class Rosenbrock<double, int>;
-
-  template class FixedStep<double>;
-
-  template class AdaptiveStep<double>;
 } // namespace Integrator
