+++ GridKit/Solver/Dynamic/Rosenbrock.hpp
@@ -12,6 +12,9 @@
 #include <GridKit/LinearAlgebra/Vector/VectorHandler.hpp>
 #include <GridKit/MemoryUtilities/MemoryUtils.hpp>
 #include <GridKit/Model/Evaluator.hpp>
+#include <GridKit/ScalarTraits.hpp>
+#include <GridKit/Solver/Dynamic/ErrorNorm.hpp>
+#include <GridKit/Solver/Dynamic/StepController.hpp>
 
 #include <resolve/Common.hpp>
 #include <resolve/SystemSolver.hpp>
@@ -20,81 +23,6 @@
 
 namespace Integrator
 {
-
-  /**
-   * @brief Define control flow for `StepController`s to be able to control the step size of a `Rosenbrock` integrator.
-   *
-   */
-  template <typename RealT>
-  struct StepControl
-  {
-    /**
-     * @brief Whether or not the step is accepted. A rejected step will cause the time step controller to discard
-     *        the next state and re-step with the new `step_size`.
-     *
-     */
-    bool  accept_;
-    /**
-     * @brief The step size the next step should take.
-     *
-     */
-    RealT step_size_;
-  };
-
-  /**
-   * @brief Interface for step size controllers. Used by `Rosenbrock` integrators to decide when to accept/reject steps and
-   *        what size each step should be.
-   *
-   * @todo It may be best to have \ref usesError() return a reference to the \ref ErrorNorm that should be used.
-   */
-  template <typename RealT>
-  class StepController
-  {
-  public:
-    /**
-     * @brief Decide the control flow for the next step, based on information gathered by the integrator about the current step.
-     *
-     * @param err The estimated error made by the current step. Only calculated if `usesError()` returns true, otherwise it is assumed this
-     *            value is not used.
-     * @param prev_step The control flow from before the current step was taken. Can be used to accurately update the step size.
-     * @param method_order The order of the method being used.
-     * @return StepControl
-     */
-    virtual StepControl<RealT> nextStep(RealT err, StepControl<RealT> prev_step, uint8_t method_order) = 0;
-    /**
-     * @brief Return whether or not the `nextStep` method implementation uses the `err` parameter. If `false`, this parameter is not calculated.
-     *
-     */
-    virtual bool               usesError() const                                                       = 0;
-  };
-
-  /**
-   * @brief Interface for error norms. Used to calculate the `err` parameter in `StepController::nextStep` based on a residual state error vector.
-   *
-   */
-  template <class ScalarT, typename IdxT>
-  class ErrorNorm
-  {
-    using State = GridKit::LinearAlgebra::Vector<ScalarT, IdxT>;
-    using RealT = typename GridKit::ScalarTraits<ScalarT>::RealT;
-
-  public:
-    /**
-     * @brief Calculate an error to be used by a step controller. Typically, an error > 1 indicates an error which does not meet tolerances, while
-     *        an error < 1 indicates an error which meets tolerances. For that reason, tolerances should be included in the calculation of the error.
-     *
-     * @param err The state error residual being measured.
-     * @param y The state that the error was calculated from. Can be used for proper relative error normalization.
-     * @param yprev The state from the previous step. Can be used for proper relative error normalization.
-     * @param handler A vector handler which can be used to facilitate vector operations.
-     * @param memspace The memory space which vector operations should be performed in/.
-     * @return The error.
-     *
-     * @todo Allow this method to fail, since it will likely involve linear algebra calls.
-     */
-    virtual RealT errorNorm(State& err, State& y, State& yprev, GridKit::LinearAlgebra::VectorHandler<ScalarT, IdxT>& handler, GridKit::memory::MemorySpace memspace) const = 0;
-  };
-
   /**
    * @brief A Rosenbrock integrator designed to simulate a `GridKit::Model::Evaluator` model.
    *        Rosenbrock integrators are best for quick medium-accuracy simulation and
@@ -605,156 +533,5 @@ namespace Integrator
 
     [[nodiscard("May fail. Check error code.")]]
     int interpDense(RealT theta);
-  };
-
-  /**
-   * @brief A simple textbook adaptive `StepController` which seeks to meet a relative and absolute tolerance
-   *        based on an error estimate.
-   *
-   */
-  template <typename RealT>
-  class AdaptiveStep : public StepController<RealT>
-  {
-    /**
-     * @brief Parameters for the step controller.
-     *
-     */
-    struct Parameters
-    {
-      /**
-       * @brief The minimum multiple by which the step size can be multiplied to obtain the new step size.
-       *        Increasing this can allow the integrator to be slightly more conservative in selecting the step size
-       *        - decreasing the number of steps taken but increasing the risk of failing the next step.
-       *
-       * @note Should be between 0 and 1.
-       *
-       */
-      RealT fac_min_   = 0.2;
-      /**
-       * @brief The maximum multiple by which the step size can be multiplied to obtain the new step size.
-       *        Decreasing this will make the integrator more conservative in selecting the step size -
-       *        increasing the number of steps taken but decreasing the risk of failing the next step.
-       *
-       * @note Should be greater than 1.
-       *
-       */
-      RealT fac_max_   = 5.0;
-      /**
-       * @brief A "fudge factor" introduced to decrease risk of failing a step. The larger the fudge factor,
-       *        the more likely steps will fail, but fewer steps will be taken.
-       *
-       * @note Should be between 0 and 1.
-       *
-       */
-      RealT fac_scale_ = 0.9;
-    } params_;
-
-  public:
-    AdaptiveStep(const Parameters& params)
-      : params_(params)
-    {
-    }
-
-    StepControl<RealT> nextStep(RealT err, StepControl<RealT> prev_step, uint8_t method_order) final;
-
-    /**
-     * @brief This controller uses error estimates.
-     *
-     * @see `nextStep()`
-     *
-     */
-    constexpr bool usesError() const final
-    {
-      return true;
-    }
-  };
-
-  /**
-   * @brief A fixed step controller which doesn't change the step size and accepts every step.
-   *        Useful if you know what time scale your simulation operates on apriori and you're
-   *        using a method without an embedded error controller.
-   *
-   *        To set the fixed size, set the `Rosenbrock::Parameters::starting_step` parameter.
-   *
-   */
-  template <typename RealT>
-  class FixedStep : public StepController<RealT>
-  {
-    StepControl<RealT> nextStep(RealT err, StepControl<RealT> prev_step, uint8_t method_order) final;
-
-    /**
-     * @brief This controller does not use error estimates.
-     *
-     * @see `nextStep()`
-     *
-     */
-    constexpr bool usesError() const final
-    {
-      return false;
-    }
-  };
-
-  /**
-   * @brief The infinity error norm, which requires the error in every component in the system
-   *        to meet tolerance.
-   *
-   */
-  template <class ScalarT, typename IdxT>
-  class InfNorm : public ErrorNorm<ScalarT, IdxT>
-  {
-    using State = GridKit::LinearAlgebra::Vector<ScalarT, IdxT>;
-    using RealT = ErrorNorm<ScalarT, IdxT>::RealT;
-
-    /**
-     * @brief A workspace for the linear algebra operations required to calculate the norm.
-     *
-     */
-    mutable struct
-    {
-      /**
-       * @brief The final vector which will have its norm taken.
-       *
-       */
-      std::unique_ptr<State> out_;
-      /**
-       * @brief The vector which will be used to scale the error in accordance with the tolerances.
-       *
-       */
-      std::unique_ptr<State> scale_;
-      /**
-       * @brief The absolute value of yprev. Used to calculate \ref scale_
-       *
-       */
-      std::unique_ptr<State> yprev_abs_;
-    } workspace_;
-
-  public:
-    /**
-     * @brief The configurable parameters of the error norm.
-     *
-     */
-    struct Parameters
-    {
-      /**
-       * @brief A vector of absolute tolerances for each component. The norm will attempt to reject errors
-       *        in each component above this tolerance.
-       *
-       */
-      std::unique_ptr<State> abs_tol_;
-
-      /**
-       * @brief The relative tolerance. The norm will attempt to reject any error larger in percentage of
-       *        the solution's maximum element than this.
-       *
-       */
-      RealT rel_tol_;
-    } params_;
-
-    InfNorm(Parameters&& params)
-      : params_(std::move(params))
-    {
-    }
-
-    RealT errorNorm(State& err, State& y, State& yprev, GridKit::LinearAlgebra::VectorHandler<ScalarT, IdxT>& handler, GridKit::memory::MemorySpace memspace) const final;
   };
 } // namespace Integrator
