+++ tests/UnitTests/Solver/Dynamic/CMakeLists.txt
@@ -11,7 +11,7 @@ endif()
 
 if(GRIDKIT_ENABLE_RESOLVE)
   add_executable(test_rosenbrock runRosenbrockTests.cpp)
-  target_link_libraries(test_rosenbrock GridKit::rosenbrock GridKit::testing)
+  target_link_libraries(test_rosenbrock GridKit::rosenbrock GridKit::testing GridKit::fixed_step ReSolve::ReSolve)
   add_test(NAME RosenbrockTest COMMAND $<TARGET_FILE:test_rosenbrock>)
 
   install(TARGETS test_rosenbrock RUNTIME DESTINATION bin)
